Modified to work with remote Selenium Hub... in theory. ##115732

+++ django_selenium/selenium_runner.py
@@ -89,30 +89,29 @@ class SeleniumTestRunner(DjangoTestSuiteRunner):
             os.environ['DISPLAY'] = settings.SELENIUM_DISPLAY
             # Start test server
             self.test_server = start_test_server(port=settings.SELENIUM_TESTSERVER_PORT)
+            if settings.SELENIUM_HOST is None:
+                # Start selenium server
+                self.selenium_server = subprocess.Popen(('java -jar %s' % settings.SELENIUM_PATH).split())
 
-            # Start selenium server
-            self.selenium_server = subprocess.Popen(('java -jar %s' % settings.SELENIUM_PATH).split())
-
-            # Waiting for server to be ready
-            if not wait_until_connectable(4444):
-                self.selenium_server.kill()
-                self.test_server.stop()
-                assert False, "selenium server does not respond"
+                # Waiting for server to be ready
+                if not wait_until_connectable(4444):
+                    self.selenium_server.kill()
+                    self.test_server.stop()
+                    assert False, "selenium server does not respond"
 
     def _stop_selenium(self):
         if self.selenium:
             # Stop selenium server
-            selenium_server = self.selenium_server
-            selenium_server.send_signal(signal.SIGINT)
-            if selenium_server.poll() is None:
-                selenium_server.kill()
-                selenium_server.wait()
-
+            if settings.SELENIUM_HOST is None:
+                selenium_server = self.selenium_server
+                selenium_server.send_signal(signal.SIGINT)
+                if selenium_server.poll() is None:
+                    selenium_server.kill()
+                    selenium_server.wait()
             # Stop test server
             self.test_server.stop()
 
     def run_tests(self, test_labels, extra_tests=None, **kwargs):
-
         self._start_selenium()
         results = super(SeleniumTestRunner, self).run_tests(test_labels, extra_tests, **kwargs)
         self._stop_selenium()
